requirements.txt
fastmcp>=0.4.0
mcp>=1.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.26.0 